+++ packages/printer-legacy/tests/unit/directive.test.js
@@ -4,7 +4,7 @@ const {
   GraphQLString,
   DirectiveLocation,
 } = require("graphql");
-const { DEFAULT_OPTIONS } = require("../../src");
+const { DEFAULT_OPTIONS } = require("../../src/printer");
 
 const {
   printCodeDirective,
