+++ packages/printer-legacy/tests/unit/input.test.js
@@ -3,7 +3,7 @@ const {
   GraphQLString,
   GraphQLInputObjectType,
 } = require("graphql");
-const { DEFAULT_OPTIONS } = require("../../src");
+const { DEFAULT_OPTIONS } = require("../../src/printer");
 
 const { printCodeInput, printInputMetadata } = require("../../src/input");
