+++ packages/printer-legacy/tests/unit/interface.test.js
@@ -3,7 +3,7 @@ const {
   GraphQLString,
   GraphQLInterfaceType,
 } = require("graphql");
-const { DEFAULT_OPTIONS } = require("../../src");
+const { DEFAULT_OPTIONS } = require("../../src/printer");
 
 const {
   printCodeInterface,
